replaced creating app menu and tray to separated files && implemented open settings modal via ipc
