--- public/lib/tray_menu_template.js
@@ -1,10 +1,0 @@
-module.exports = (win) => {
-  return [
-    { label: 'Honey Bitfinex' },
-    { type: 'separator' },
-    {
-      label: 'Hide/show application',
-      click: () => (win.isVisible() ? win.hide() : win.show()),
-    },
-  ]
-}
